+++ spec/safedep/gemfile_lock_spec.rb
@@ -76,6 +76,12 @@ module Safedep
       it 'returns an array of the specifications' do
         expect(dep_names).to include('rake', 'rspec', 'rubocop')
       end
+
+      it 'includes "bundler" dependency' do
+        dependency = lockfile.find_dependency('bundler')
+        expect(dependency.name).to eq('bundler')
+        expect(dependency.version).to eq(Bundler::VERSION)
+      end
     end
   end
 end