+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Development
 
+* Fix an error when a `Gemfile` includes `bundler` dependency without version specifier.
+
 ## v0.0.1
 
 * Initial release.